requirements.txt
alembic==1.17.1
aniso8601==10.0.1
appnope==0.1.4
asttokens==3.0.0
attrs==25.4.0
blinker==1.9.0
certifi==2025.10.5
click==8.3.0
comm==0.2.3
debugpy==1.8.17
decorator==5.2.1
Deprecated==1.3.1
dominate==2.9.1
executing==2.2.1
Flask==3.1.2
Flask-Admin==2.0.1
Flask-Bootstrap==3.3.7.1
flask-cors==6.0.1
Flask-Limiter==4.0.0
flask-marshmallow==1.3.0
Flask-Migrate==4.1.0
flask-restx==1.3.2
Flask-SQLAlchemy==3.1.1
geopandas==1.1.1
importlib_resources==6.5.2
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
limits==5.6.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==4.1.0
matplotlib-inline==0.2.1
mdurl==0.1.2
nest-asyncio==1.6.0
numpy==2.3.4
ordered-set==4.1.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
psutil==7.1.3
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyogrio==0.11.1
pyproj==3.7.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pyzmq==27.1.0
referencing==0.37.0
rich==14.2.0
rpds-py==0.28.0
shapely==2.1.2
six==1.17.0
SQLAlchemy==2.0.44
stack-data==0.6.3
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
visitor==0.1.3
waitress==3.0.2
wcwidth==0.2.14
Werkzeug==3.1.3
wrapt==2.0.1
WTForms==3.2.1
